Updated angularjs template to handle multiple repositories.
This required the Python script to output a JSON file containing a list of repositories (and paths to the data).

I also updated the template so that comments could be linked to, and ensured that scrolling to comments worked correctly both when clickling on a link on the same page and via a direct address bar link. This was tricky since angularjs uses # for routing too!