hg2git: fix splitting logic in two methods
get_git_log and get_hg_hashes_from_git both use random UUID's as delimiters
in formatted git log output and split the data on a newline followed by
the specific UUID.
This fails for the first line in the formatted log.
As a quick workaround, I've prepended a newline to the log data so it
should split correctly.