Fixed bug with repository creation on GitHub
If a repository existed at the correct location (owner/repo_name) and the repository was moved to a new owner (either via transfer or by renaming an organisation) then this tool would follow the redirect and not recreate the repository at the expected location). We were catching this correctly (recreating the repository) if just the repository was renamed within the owner, but we were not previously ensuring the owner was as expected. This is now fixed.