hg2git: encode git logs as ascii for consistency
In 310e36e29d401, the hg log output is encoded as ascii to match the output
of Github's source import tool. This can cause a mismatch when using a
different source conversion tool that doesn't encode git logs to ascii.
For consistency, the hg2git modules encode both hg and git logs as ascii
before comparing them for matches.