Ignore refs/branches to reduce export size / time
I have exported some very large repositories and noticed
that the bitbucket_data_raw/repositories/*/*/commits folder
is by far the largest. Upon inspection, it appeared to be
dominated by multi-page json files for each branch.
This change adds an ignore rule for refs/branches URL's,
which substantially reduces the size of the commits folder
and the time it takes to export.
As a side-effect, it also excludes the refs/branches*.json
files and refs/branches/ folder, but that is not currently used
by the github-pages and seems like a worthwhile trade-off.